refactor(profiling): extract thread_utils module
