+++ profiling/src/profiling.rs
@@ -1,3 +1,4 @@
+use super::thread_utils;
 use crate::bindings::{
     datadog_php_profiling_get_profiling_context, zend_execute_data, zend_function, zend_string,
     ZEND_USER_FUNCTION,
@@ -9,13 +10,11 @@ use crossbeam_channel::{select, Receiver, Sender, TrySendError};
 use datadog_profiling::exporter::{Endpoint, File, Tag};
 use datadog_profiling::profile;
 use datadog_profiling::profile::api::{Function, Line, Location, Period, Sample};
-use libc::sched_yield;
 use log::{debug, info, trace, warn};
 use std::borrow::{Borrow, Cow};
 use std::collections::HashMap;
 use std::hash::Hash;
 use std::intrinsics::transmute;
-use std::mem::MaybeUninit;
 use std::str;
 use std::str::Utf8Error;
 use std::sync::atomic::{AtomicBool, AtomicU32, Ordering};
@@ -621,46 +620,6 @@ impl Uploader {
 }
 
 impl Profiler {
-    /// Spawns a thread and masks off the signals that the Zend Engine uses.
-    fn spawn<F, T>(name: &str, f: F) -> JoinHandle<T>
-    where
-        F: FnOnce() -> T + Send + 'static,
-        T: Send + 'static,
-    {
-        let result = std::thread::Builder::new()
-            .name(name.to_string())
-            .spawn(move || {
-                /* Thread must not handle signals intended for PHP threads.
-                 * See Zend/zend_signal.c for which signals it registers.
-                 */
-                unsafe {
-                    let mut sigset_mem = MaybeUninit::uninit();
-                    let sigset = sigset_mem.as_mut_ptr();
-                    libc::sigemptyset(sigset);
-
-                    const SIGNALS: [libc::c_int; 6] = [
-                        libc::SIGPROF, // todo: SIGALRM on __CYGWIN__/__PHASE__
-                        libc::SIGHUP,
-                        libc::SIGINT,
-                        libc::SIGTERM,
-                        libc::SIGUSR1,
-                        libc::SIGUSR2,
-                    ];
-
-                    for signal in SIGNALS {
-                        libc::sigaddset(sigset, signal);
-                    }
-                    libc::pthread_sigmask(libc::SIG_BLOCK, sigset, std::ptr::null_mut());
-                }
-                f()
-            });
-
-        match result {
-            Ok(handle) => handle,
-            Err(err) => panic!("Failed to spawn thread {name}: {err}"),
-        }
-    }
-
     pub fn new() -> Self {
         let fork_barrier = Arc::new(Barrier::new(3));
         let (fork_sender0, fork_receiver0) = crossbeam_channel::bounded(1);
@@ -688,8 +647,8 @@ impl Profiler {
             fork_senders: [fork_sender0, fork_sender1],
             vm_interrupts,
             message_sender,
-            time_collector_handle: Self::spawn("ddprof_time", move || time_collector.run()),
-            uploader_handle: Self::spawn("ddprof_upload", move || uploader.run()),
+            time_collector_handle: thread_utils::spawn("ddprof_time", move || time_collector.run()),
+            uploader_handle: thread_utils::spawn("ddprof_upload", move || uploader.run()),
         }
     }
 
@@ -740,33 +699,6 @@ impl Profiler {
             .try_send(ProfilerMessage::LocalRootSpanResource(message))
     }
 
-    /// Waits for the handle to be finished. If finished, it will join the
-    /// handle. Otherwise, it will leak the handle.
-    /// # Panics
-    /// Panics if the thread being joined has panic'd.
-    fn join_timeout(handle: JoinHandle<()>, timeout: Duration, impact: &str) {
-        // After notifying the other threads, it's likely they'll need some
-        // time to respond adequately. Joining on the JoinHandle is supposed
-        // to be the correct way to do this, but we've observed this can
-        // panic:
-        // https://github.com/DataDog/dd-trace-php/issues/1919
-        // Thus far, we have not been able to reproduce it and address the
-        // root cause. So, for now, mitigate it instead with a busy loop.
-        let start = Instant::now();
-        while !handle.is_finished() {
-            unsafe { sched_yield() };
-            if start.elapsed() >= timeout {
-                let name = handle.thread().name().unwrap_or("{unknown}");
-                warn!("Timeout of {timeout:?} reached when joining thread '{name}'. {impact}");
-                return;
-            }
-        }
-
-        if let Err(err) = handle.join() {
-            std::panic::resume_unwind(err)
-        }
-    }
-
     pub fn stop(self) {
         debug!("Stopping profiler.");
         match self.message_sender.send(ProfilerMessage::Cancel) {
@@ -775,7 +707,7 @@ impl Profiler {
         }
 
         let timeout = Duration::from_secs(2);
-        Self::join_timeout(
+        thread_utils::join_timeout(
             self.time_collector_handle,
             timeout,
             "Recent samples may be lost.",
@@ -784,7 +716,7 @@ impl Profiler {
         // Wait for the time_collector to join, since that will drop
         // the sender half of the channel that the uploader is
         // holding, allowing it to finish.
-        Self::join_timeout(
+        thread_utils::join_timeout(
             self.uploader_handle,
             timeout,
             "Recent samples are most likely lost.",
